% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cation2Cation.R
\name{Cation2Cation}
\alias{Cation2Cation}
\title{Converting atomic proportions into atomic proportions with different normalization methods and/or element choices}
\usage{
Cation2Cation(
  table,
  ElementList,
  NormMode = "Oxygen",
  NormValue = 24,
  padding = FALSE
)
}
\arguments{
\item{table}{(dataframe) Atomic proportion data}

\item{ElementList}{(dataframe) Information of considered elements, cation/oxygen numbers per unit formula.
Use MFRC::ElementList_all or its subset (e.g., MFRC::ElementList_MnNCKFMASTCr, MFRC::ElementList_NCFMAS).}

\item{NormMode}{(chr) Set "Oxygen" or "Cation" (default = "Oxygen")}

\item{NormValue}{(num) Normalization value (default = 24)}

\item{padding}{(logical) TRUE for padding concentrations for element absent in the input data (default = FALSE)}
}
\description{
\code{Cation2Cation} converts atomic proportions into atomic proportions with different normalization methods and/or element choices.
}
\examples{
Mass2Cation(MFRC::testwtpc, MFRC::ElementList_NCFMAS, NormMode = "Oxygen", NormValue = 24, padding = FALSE) |>
          Cation2Cation(MFRC::ElementList_MnNCKFMASTCr, NormMode = "Cation", NormValue = 4, padding = TRUE)


}
