% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cation2MassPc.R
\name{Cation2MassPc}
\alias{Cation2MassPc}
\title{Converting atomic proportions into wt\%}
\usage{
Cation2MassPc(table, ElementList, NormValue = 100, padding = FALSE)
}
\arguments{
\item{table}{(dataframe) Atomic proportion data}

\item{ElementList}{(dataframe) Information of considered elements, cation/oxygen numbers per unit formula.
Use MFRC::ElementList_all or its subset (e.g., MFRC::ElementList_MnNCKFMASTCr, MFRC::ElementList_NCFMAS).}

\item{NormValue}{(num) Normalization value for total (default = 100)}

\item{padding}{(logical) TRUE for padding concentrations for element absent in the input data (default = FALSE)}
}
\description{
\code{Cation2MassPc} converts atomic proportions into wt\%.
}
\examples{
Mass2Cation(MFRC::testwtpc, MFRC::ElementList_NCFMAS, NormMode = "Oxygen", NormValue = 24, padding = FALSE) |>
          Cation2MassPc(MFRC::ElementList_MnNCKFMASTCr, NormValue = 100, padding = TRUE)


}
