% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ElementList_all.R
\docType{data}
\name{ElementList_all}
\alias{ElementList_all}
\title{Information of considered elements, cation/oxygen numbers per unit formula (all elements).}
\format{
dataframe
}
\usage{
ElementList_all
}
\description{
Information of considered elements, cation/oxygen numbers per unit formula (all elements).
}
\keyword{datasets}
