% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ElementList_MnNCKFMASTCr.R
\docType{data}
\name{ElementList_MnNCKFMASTCr}
\alias{ElementList_MnNCKFMASTCr}
\title{Information of considered elements, cation/oxygen numbers per unit formula (MnNCKFMASTCr).}
\format{
dataframe
}
\usage{
ElementList_MnNCKFMASTCr
}
\description{
Information of considered elements, cation/oxygen numbers per unit formula (MnNCKFMASTCr).
}
\keyword{datasets}
