% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ElementList_NCFMAS.R
\docType{data}
\name{ElementList_NCFMAS}
\alias{ElementList_NCFMAS}
\title{Information of considered elements, cation/oxygen numbers per unit formula (NCFMAS).}
\format{
dataframe
}
\usage{
ElementList_NCFMAS
}
\description{
Information of considered elements, cation/oxygen numbers per unit formula (NCFMAS).
}
\keyword{datasets}
