% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mass2Cation.R
\name{Mass2Cation}
\alias{Mass2Cation}
\title{Converting wt\% into atomic proportions}
\usage{
Mass2Cation(
  table,
  ElementList,
  NormMode = "Oxygen",
  NormValue = 24,
  padding = FALSE
)
}
\arguments{
\item{table}{(dataframe) wt\% data}

\item{ElementList}{(dataframe) Information of considered elements, cation/oxygen numbers per unit formula.
Use MFRC::ElementList_all or its subset (e.g., MFRC::ElementList_MnNCKFMASTCr, MFRC::ElementList_NCFMAS).}

\item{NormMode}{(chr) Set "Oxygen" or "Cation" (default = "Oxygen")}

\item{NormValue}{(num) Normalization value (default = 24)}

\item{padding}{(logical) TRUE for padding concentrations for element absent in the input wt\% data (default = FALSE)}
}
\description{
\code{Mass2Cation} converts wt\% into atomic proportions.
}
\examples{
table <- MFRC::testwtpc
ElementList <- MFRC::ElementList_MnNCKFMASTCr
Result1 <- Mass2Cation(table, ElementList, NormMode = "Oxygen", NormValue = 24, padding = FALSE)
}
