% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MassList.R
\docType{data}
\name{MassList}
\alias{MassList}
\title{List of molecular weights (copied from MinPlot)}
\format{
dataframe
}
\usage{
MassList
}
\description{
List of molecular weights (copied from MinPlot)
}
\keyword{datasets}
