% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Omp_AlT_Omit.R
\name{Omp_AlT_Omit}
\alias{Omp_AlT_Omit}
\title{Recalculation for omphacite (NCFMAS, Al_T is omitted, Ca + Mg + Fe in M2 = Mg + Fe in M1, Mg/Fe equipartition)}
\usage{
Omp_AlT_Omit(CationTable)
}
\arguments{
\item{CationTable}{(dataframe) Atomic proportion data}
}
\description{
\code{Omp_AlT_Omit}: For omphacite (NCFMAS, Al_T is omitted)
}
\examples{
Mass2Cation(MFRC::testwtpc, MFRC::ElementList_MnNCKFMASTCr) |> Omp_AlT_Omit()

}
