% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pyroxene_MinPlot.R
\name{Pyroxene_MinPlot}
\alias{Pyroxene_MinPlot}
\title{Recalculation for pyroxene (method from MinPlot)}
\usage{
Pyroxene_MinPlot(CationTable)
}
\arguments{
\item{CationTable}{(dataframe) Atomic proportion data}
}
\description{
\code{Pyroxene_MinPlot} calculates site fractions for pyroxene with the method used in MinPlot (Walters, 2022).
}
\examples{
Mass2Cation(MFRC::testwtpc, MFRC::ElementList_MnNCKFMASTCr) |> Pyroxene_MinPlot()
}
\references{
Walters, J. B. (2022). A mineral formula recalculation and plotting program for electron probe microanalysis.
Mineralogia, 53(1), 51-66.
}
